created home component
refactored to have routes
