+++ src/app/app.module.ts
@@ -3,6 +3,10 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 
+import {routing} from './app.routes';
+
+import {HomeModule} from './home/home.module';
+
 import { AppComponent } from './app.component';
 
 @NgModule({
@@ -12,7 +16,9 @@ import { AppComponent } from './app.component';
   imports: [
     BrowserModule,
     FormsModule,
-    HttpModule
+    HttpModule,
+    routing,
+    HomeModule
   ],
   providers: [],
   bootstrap: [AppComponent]
