+++ src/app/home/home.module.ts
@@ -1,0 +1,16 @@
+import { NgModule} from '@angular/core';
+import {HomeComponent} from './home.component';
+import { BrowserModule } from '@angular/platform-browser';
+import { FormsModule } from '@angular/forms';
+import { HttpModule } from '@angular/http';
+
+@NgModule({
+    imports: [
+        BrowserModule,
+        FormsModule,
+        HttpModule],
+    declarations: [HomeComponent],
+    exports: [HomeComponent],
+    entryComponents: []
+})
+export class HomeModule { }
