+++ src/app/home/home.routes.ts
@@ -1,0 +1,14 @@
+import { Route } from '@angular/router';
+import { HomeComponent} from './home.component';
+
+export const HomeRoutes: Route[] = [
+    {
+        path: '',
+        redirectTo: 'home',
+        pathMatch: 'full'
+    },
+    {
+        path: 'home',
+        component: HomeComponent
+    }
+];
