+++ src/app/home/index.ts
@@ -1,0 +1,2 @@
+export * from './home.component';
+export * from './home.routes';
